+++ flowmachine/flowmachine/features/subscriber/__init__.py
@@ -63,6 +63,8 @@ from .meaningful_locations import (
     MeaningfulLocationsOD,
 )
 
+from .topup_amount import TopUpAmount
+from .topup_balance import TopUpBalance
 from .mds_volume import MDSVolume
 from .event_type_proportion import ProportionEventType
 
@@ -112,6 +114,8 @@ __all__ = [
     "LocationEntropy",
     "ContactEntropy",
     "DistanceCounterparts",
+    "TopUpAmount",
+    "TopUpBalance",
     "MDSVolume",
     "ContactReciprocal",
     "ProportionEventReciprocal",